refactor: amount 원시값 포장
