+++ src/main/kotlin/lotto/model/Amount.kt
@@ -1,0 +1,32 @@
+package lotto.model
+
+@JvmInline
+value class Amount(
+    private val amount: Int,
+) {
+    init {
+        validateAmountMinimumRange()
+        validateAmountUnit()
+    }
+
+    fun isAffordable(manualQuantity: Int): Boolean = amount - manualQuantity * LOTTO_EACH_AMOUNT >= 0
+
+    fun getAutoLottoQuantity(manualQuantity: Int): Int = amount / LOTTO_EACH_AMOUNT - manualQuantity
+
+    private fun validateAmountMinimumRange() {
+        require(amount > LOTTO_MIN_AMOUNT) {
+            "[ERROR] ${LOTTO_MIN_AMOUNT}원 이상의 금액으로 입력해 주세요. 입력값: $amount"
+        }
+    }
+
+    private fun validateAmountUnit() {
+        require(amount % LOTTO_EACH_AMOUNT == 0) {
+            "[ERROR] ${LOTTO_EACH_AMOUNT}원 단위의 금액으로 입력해 주세요. 입력값: $amount"
+        }
+    }
+
+    companion object {
+        private const val LOTTO_MIN_AMOUNT = 0
+        private const val LOTTO_EACH_AMOUNT = 1000
+    }
+}
