+++ src/main/kotlin/lotto/model/LottoMarket.kt
@@ -1,16 +1,14 @@
 package lotto.model
 
 class LottoMarket(
-    private val amount: Int,
+    private val amount: Amount,
     private val manualQuantity: Int,
     private val manualLottoMachine: ManualLottoMachine = ManualLottoMachine(),
     private val autoLottoMachine: AutoLottoMachine = AutoLottoMachine(),
 ) {
-    val autoLottoQuantity: Int = amount / LOTTO_EACH_AMOUNT - manualQuantity
+    val autoLottoQuantity: Int = amount.getAutoLottoQuantity(manualQuantity)
 
     init {
-        validateAmountMinimumRange()
-        validateAmountUnit()
         validateManualLottoQuantity()
     }
 
@@ -26,28 +24,13 @@ class LottoMarket(
         )
     }
 
-    private fun validateAmountMinimumRange() {
-        require(amount > LOTTO_MIN_AMOUNT) {
-            "[ERROR] ${LOTTO_MIN_AMOUNT}원 이상의 금액으로 입력해 주세요. 입력값: $amount"
-        }
-    }
-
-    private fun validateAmountUnit() {
-        require(amount % LOTTO_EACH_AMOUNT == 0) {
-            "[ERROR] ${LOTTO_EACH_AMOUNT}원 단위의 금액으로 입력해 주세요. 입력값: $amount"
-        }
-    }
-
     private fun validateManualLottoQuantity() {
-        require(amount - manualQuantity * LOTTO_EACH_AMOUNT >= 0) {
+        require(amount.isAffordable(manualQuantity)) {
             "[ERROR] 낸 금액보다 많은 수동 로또를 살 수 없습니다. 금액: $amount, 수동 로또: ${manualQuantity}장"
         }
     }
 
     companion object {
         const val EMPTY_LOTTO_QUANTITY = 0
-
-        private const val LOTTO_MIN_AMOUNT = 0
-        private const val LOTTO_EACH_AMOUNT = 1000
     }
 }
